--- bootstrap/leg/leg.go
@@ -1,1 +1,0 @@
-../../leg.go
