--- grammars/fexl/fexl.peg
@@ -1,35 +1,0 @@
-package main
-
-type Fexl Peg {
- 
-}
-
-Fexl <- ws Expression+ Input? !.
-
-Input <- '\\\\' .*
-
-Expression <- Comment / ';' ws Expression* / Definition / Argument / Term
-
-Comment <- '#' (![\n\r] .)* ws
-
-Definition <- '\\' Symbol '=' ws Term / Recursive
-
-Recursive <- '\\' Symbol '==' ws Term
-
-Argument <- '\\' Symbol
-
-Term <- open Expression+ close / Symbol
-
-Symbol <- (String / (![ \t\n\r\\()"~;=] .)+) ws
-
-String <- '"' (!'"' .)* '"' / Complex
-
-Complex <- tilde '@' (!'@' .)* '@'
-
-tilde <- '~'
-
-open <- '(' ws
-
-close <- ')' ws
-
-ws <- [ \t\n\r]*
