--- grammars/long_test/long.peg
@@ -1,11 +1,0 @@
-# Copyright 2010 The Go Authors. All rights reserved.
-# Use of this source code is governed by a BSD-style
-# license that can be found in the LICENSE file.
-
-package main
-
-type Long Peg {
-
-}
-
-String <- '\"' (!'\"' .)* '\"' !.
