+++ src/bootstrap/leg/leg.go
@@ -1,0 +1,1 @@
+../../leg/leg.go
